+++ IoT/physical.py
@@ -1,7 +1,5 @@
+print("sensors and acutators")
 import time
-
-print("Sensor and Actuator")
-
 import serial.tools.list_ports
 
 def getPort():
@@ -14,24 +12,15 @@ def getPort():
         if "USB Serial" in strPort:
             splitPort = strPort.split(" ")
             commPort = (splitPort[0])
-    return commPort
+        return commPort
 
 portName = getPort()
-
+print(portName)
 if portName != "None":
-    ser = serial.Serial(port = portName, baudrate = 9600)
+    ser = serial.Serial(port=portName, baudrate=9600)
 
-# ser = serial.Serial(port = portName, baudrate = 9600)
-# print(ser)
-
-relay1_ON = [0, 6, 0, 0, 0, 255, 200, 91] # crc 16-bit checksum
+relay1_ON  = [0, 6, 0, 0, 0, 255, 200, 91]
 relay1_OFF = [0, 6, 0, 0, 0, 0, 136, 27]
-
-relay2_ON = [15, 6, 0, 0 ,0, 255, 200, 164]
-relay2_OFF = [15, 6, 0, 0, 0, 0, 136, 228]
-
-# def turn_ON_or_OFF(myList = [], *args):
-#     ser.write(myList)
 
 def setDevice1(state):
     if state == True:
@@ -39,19 +28,43 @@ def setDevice1(state):
     else:
         ser.write(relay1_OFF)
 
+relay2_ON  = [15, 6, 0, 0, 0, 255, 200, 164]
+relay2_OFF = [15, 6, 0, 0, 0, 0, 136, 228]
+
 def setDevice2(state):
     if state == True:
         ser.write(relay2_ON)
     else:
         ser.write(relay2_OFF)
 
-#
+def serial_read_data(ser):
+    bytesToRead = ser.inWaiting()
+    if bytesToRead > 0:
+        out = ser.read(bytesToRead)
+        data_array = [b for b in out]
+        print(data_array)
+        if len(data_array) >= 7:
+            array_size = len(data_array)
+            value = data_array[array_size - 4] * 256 + data_array[array_size - 3]
+            return value
+        else:
+            return -1
+    return 0
+
+soil_temperature =[1, 3, 0, 6, 0, 1, 100, 11]
+def readTemperature():
+    serial_read_data(ser)
+    ser.write(soil_temperature)
+    time.sleep(5)
+    return serial_read_data(ser)
+
+soil_moisture = [1, 3, 0, 7, 0, 1, 53, 203]
+def readMoisture():
+    serial_read_data(ser)
+    ser.write(soil_moisture)
+    time.sleep(5)
+    return serial_read_data(ser)
+
 # while True:
-#     ser.write(relay1_ON)
-#     time.sleep(2)
-#     ser.write(relay1_OFF)
-#     time.sleep(10)
-#
-# # 255 all 1 => on
-# # 200,91 => error code
-# # CRC 16 bit, crc table, check sum
+#     print(readTemperature())
+#     print(readMoisture())
